+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Issue #3: Stop at platform for loading
 - Issue #4: Immutable geometry, routes, trains
+- Issue #6: Lock signals when train changes section
 - Issue #7: Approaching train state
 - Issue #8: Braking train state
 - Issue #9: Exiting train state